+++ ron-test/LwwStruct/Types.hs
@@ -33,6 +33,8 @@ module LwwStruct.Types (
     tfpatInnerField_assign,
     tfpatInnerField_read,
     tfpatInnerField_zoom,
+    TestEnum1,
+    TestEnum2 (..),
 ) where
 
 import           RON.Data (Replicated, ReplicatedAsPayload, encoding,
@@ -67,6 +69,9 @@ instance ReplicatedAsPayload TestOpaque where
     (struct_lww TestFieldPrefixAndTitle
         #haskell {field_prefix "tfpat", field_case title}
         innerField Integer)
+
+    (enum TestEnum1)
+    (enum TestEnum2 TE2_Item1 TE2_Item2)
 |]
 
 deriving instance Eq   Example1